rpath: Fix leaking corrupted IFS into split package functions
When assigning a variable and then running a command, the variable is
local to the command... but here we just assign two variables, so that
isn't going to work.

In split packages, the second (and subsequent) package functions are run
after the first package is linted, by which time our very nonstandard
IFS is there to confuse find commands etc.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>